Be more strict about types. Apparently some versions of PyQt5 don't automatically convert between int and float, which is something some parts of Reggie Next make use of.

+++ background.py
@@ -266,7 +266,7 @@ class BGTab(QtWidgets.QWidget):
                     filename = os.path.join('reggiedata', ['bga', 'bgb'][slot_id], 'no_preview.png')
 
                 pix = QtGui.QPixmap(filename)
-                pix = pix.scaled(pix.width() * scale, pix.height() * scale)
+                pix = pix.scaled(int(pix.width() * scale), int(pix.height() * scale))
                 previews[slot_id][box_num].setPixmap(pix)
 
             # Alignment mode
